Add float32 support
