+++ realm/realm-library/src/syncIntegrationTest/kotlin/io/realm/SyncSessionTests.kt
@@ -16,7 +16,6 @@ import io.realm.mongodb.*
 import io.realm.mongodb.sync.*
 import io.realm.rule.BlockingLooperThread
 import io.realm.util.ResourceContainer
-import io.realm.util.assertFailsWithErrorCode
 import io.realm.util.assertFailsWithMessage
 import org.bson.BsonInt32
 import org.bson.BsonInt64
@@ -255,26 +254,6 @@ class SyncSessionTests {
             realm.syncSession.downloadAllServerChanges()
             realm.refresh()
             assertEquals(1, realm.where(SyncAllTypes::class.java).count())
-        }
-    }
-
-    // TODO This test is only for tracking failure when uploading SyncAllTypes including float field.
-    //  Once this test fails (meaning that the full schema can be uploaded) the test can be removed
-    //  and we can include the float field in SyncAllTypes
-    @Test
-    fun uploadDownloadAllChangesWithFloatFails() {
-        val config = configFactory
-                .createSyncConfigurationBuilder(user, syncConfiguration.partitionValue)
-                .testSchema(SyncAllTypesWithFloat::class.java, SyncAllTypes::class.java, SyncDog::class.java, SyncPerson::class.java)
-                .build()
-
-        Realm.getInstance(config).use { realm ->
-            realm.executeTransaction {
-                realm.createObject(SyncAllTypesWithFloat::class.java, ObjectId())
-            }
-            assertFailsWithErrorCode(ErrorCode.BAD_CHANGESET) {
-                realm.syncSession.uploadAllLocalChanges()
-            }
         }
     }
